Tests/LibWeb/Text/expected/HTML/Location-search-setter.txt

location.search set to ?foo=bar
